pyproject.toml
[tool.poetry]
name = "alphaterminal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.2"
aiohttp = "^3.8.1"
numpy = "^1.23.2"
aiofiles = "^0.8.0"
linkedin-api = "^1.1.0"
plotly = "^5.10.0"
chart-studio = "^1.1.0"
SQLAlchemy = "^1.4.40"
selenium = "^4.4.3"
email-validator = "^2.0.0"
openai = "^0.27.2"
fastapi = "^0.95.0"
gunicorn = "^20.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
loguru = "^0.7.0"
bcrypt = "4.0.1"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
scikit-learn = "^1.2.2"
python-multipart = "^0.0.6"
pytest = "^7.3.1"
pillow = "^9.5.0"
notifiers = "^1.3.3"
pydantic-duality = "^1.1.0"
pydantic = "^1.10.8"
pytest-cov = "^4.1.0"
freezegun = "^1.2.2"
python-dotenv = "^1.0.0"
pymongo = "^4.5.0"
ordered-enum = "^0.0.8"
fastapi-cache2 = "^0.2.1"
celery = "^5.3.6"
api-analytics = "^1.2.2"

[tool.poetry.group.crm_fastapi.dependencies]
uvicorn = "^0.22.0"
arbm_core = {git = "git@github.com:iteriter/arbm_core.git", branch = "main"}

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
autoflake = "^1.5.1"
pyupgrade = "^2.37.3"
bandit = "^1.7.4"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(message)s"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
skip_gitignore = true
skip_glob = ["**/tests/*"]
line_length = 120
lines_between_sections = 1

[tool.black]
line-length = 120
target_version = ['py310']

[tool.bandit]
exclude_dirs = ["**/tests/*"]
skips = ["B104"]
